+++ service/src/test/kotlin/io/provenance/explorer/service/TokenServiceTest.kt
@@ -1,11 +1,11 @@
 package io.provenance.explorer.service
 
-import io.provenance.explorer.domain.models.OsmosisHistoricalPrice
+import io.provenance.explorer.domain.models.HistoricalPrice
+import io.provenance.explorer.grpc.flow.FlowApiGrpcClient
 import io.provenance.explorer.grpc.v1.AccountGrpcClient
+import io.provenance.explorer.service.pricing.fetchers.HistoricalPriceFetcherFactory
 import kotlinx.coroutines.runBlocking
 import org.joda.time.DateTime
-import org.joda.time.DateTimeZone
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
@@ -14,51 +14,77 @@ import java.net.URI
 class TokenServiceTest {
 
     private lateinit var accountClient: AccountGrpcClient
+    private lateinit var flowApiGrpcClient: FlowApiGrpcClient
+    private lateinit var historicalPriceFetcherFactory: HistoricalPriceFetcherFactory
     private lateinit var tokenService: TokenService
 
     @BeforeEach
     fun setUp() {
-        accountClient = AccountGrpcClient(URI("https://www.google.com"))
-        tokenService = TokenService(accountClient)
+        accountClient = AccountGrpcClient(URI("http://localhost:26657"))
+        flowApiGrpcClient = FlowApiGrpcClient(URI("http://localhost:50051"))
+        historicalPriceFetcherFactory = HistoricalPriceFetcherFactory(flowApiGrpcClient) // Set the factory as needed
+        tokenService = TokenService(accountClient, historicalPriceFetcherFactory)
     }
 
     @Test
     @Disabled("Test was used to manually call the endpoint")
-    fun `test fetchOsmosisData and print results`() = runBlocking {
-        val fromDate = DateTime.parse("2024-05-08")
+    fun `test fetchHistoricalPriceData`() = runBlocking {
+        val fromDate = DateTime.now().minusDays(7)
 
-        val result: List<OsmosisHistoricalPrice> = tokenService.fetchOsmosisData(fromDate)
+        val result: List<HistoricalPrice> = tokenService.fetchHistoricalPriceData(fromDate)
 
         result.forEach {
-            println("Time: ${DateTime(it.time * 1000)}, Open: ${it.open}, High: ${it.high}, Low: ${it.low}, Close: ${it.close}, Volume: ${it.volume}")
+            println("Time: ${DateTime(it.time)}, Open: ${it.open}, High: ${it.high}, Low: ${it.low}, Close: ${it.close}, Volume: ${it.volume}")
         }
+
+        assert(result.isNotEmpty()) { "Expected non-empty list of HistoricalPrice" }
     }
 
     @Test
-    fun `test determineTimeFrame`() {
-        val now = DateTime.now(DateTimeZone.UTC)
+    @Disabled("Test was used to manually call the endpoint")
+    fun `test fetchHistoricalPriceData and process `() = runBlocking {
+        val fromDate = DateTime.now().minusDays(1)
 
-        val fromDate1 = now.minusDays(10)
-        val timeFrame1 = tokenService.determineTimeFrame(fromDate1)
-        assertEquals(TokenService.TimeFrame.FIVE_MINUTES, timeFrame1)
+        val result: List<HistoricalPrice> = tokenService.fetchHistoricalPriceData(fromDate)
 
-        val fromDate2 = now.minusDays(30)
-        val timeFrame2 = tokenService.determineTimeFrame(fromDate2)
-        assertEquals(TokenService.TimeFrame.TWO_HOURS, timeFrame2)
+        tokenService.processLatestTokenData(result, fromDate)
 
-        val fromDate3 = now.minusDays(90)
-        val timeFrame3 = tokenService.determineTimeFrame(fromDate3)
-        assertEquals(TokenService.TimeFrame.ONE_DAY, timeFrame3)
+        result.forEach {
+            println("Time: ${DateTime(it.time)}, Open: ${it.open}, High: ${it.high}, Low: ${it.low}, Close: ${it.close}, Volume: ${it.volume}")
+        }
+
+        assert(result.isNotEmpty()) { "Expected non-empty list of HistoricalPrice" }
     }
 
     @Test
-    fun `test buildInputQuery`() {
-        val now = DateTime.now(DateTimeZone.UTC)
+    @Disabled("Test was used to manually call the endpoint")
+    fun `test fetchLegacyHistoricalPriceData`() = runBlocking {
+        val fromDate = DateTime.now().minusDays(7)
 
-        val fromDate1 = now.minusDays(10)
-        val timeFrame1 = TokenService.TimeFrame.FIVE_MINUTES
-        val inputQuery1 = tokenService.buildInputQuery(fromDate1, timeFrame1)
-        val expectedQuery1 = """%7B%22json%22%3A%7B%22coinDenom%22%3A%22ibc%2FCE5BFF1D9BADA03BB5CCA5F56939392A761B53A10FBD03B37506669C3218D3B2%22%2C%22coinMinimalDenom%22%3A%22ibc%2FCE5BFF1D9BADA03BB5CCA5F56939392A761B53A10FBD03B37506669C3218D3B2%22%2C%22timeFrame%22%3A%7B%22custom%22%3A%7B%22timeFrame%22%3A5%2C%22numRecentFrames%22%3A2880%7D%7D%7D%7D"""
-        assertEquals(expectedQuery1, inputQuery1)
+        val result: List<HistoricalPrice> = tokenService.fetchLegacyHistoricalPriceData(fromDate)
+
+        result.forEach {
+            println("Time: ${DateTime(it.time)}, Open: ${it.open}, High: ${it.high}, Low: ${it.low}, Close: ${it.close}, Volume: ${it.volume}")
+        }
+
+        assert(result.isNotEmpty()) { "Expected non-empty list of HistoricalPrice" }
+    }
+
+    @Test
+    @Disabled("Test was used to manually call the endpoint")
+    fun `test getTokenDistributionStats`() {
+        val result = tokenService.getTokenDistributionStats()
+        println(result)
+        assert(result.isNotEmpty()) { "Expected non-empty token distribution stats" }
+    }
+
+    @Test
+    @Disabled("Test was used to manually call the endpoint")
+    fun `test getTokenBreakdown`() = runBlocking {
+        val result = tokenService.getTokenBreakdown()
+        println("Max supply: ${result.maxSupply}")
+        println("Community pool supply: ${result.communityPool}")
+        println("Bonded supply: ${result.bonded}")
+        println("Burned supply: ${result.burned}")
     }
 }